add sequence_counter.cpp and sequence_counter.hpp
